fix: codec table and dag-json and dag-cbor
Codec table was missing dag-json and it had invalid code for dag-cbor.
It also had invalid string representation of dag-pb -- it was using
'protobuf' which is a totally different code.

I fixed those bugs, and added generic variants for Protobuf, JSON and
CBOR, so go-cid finally match the source of truth at
https://github.com/multiformats/multicodec/blob/master/table.csv